+++ packages/antd-management-fast-design-playground/src/components/displayer/index.jsx
@@ -1,1 +1,2 @@
+export * from './DataDisplayer';
 export * from './SchemaDisplayer';