+++ packages/antd-management-fast-design-react/src/components/designable/react/locales/global.js
@@ -1,12 +1,12 @@
 const locales = {
   'zh-CN': {
-    save: '本地暂存',
+    save: '暂存设计',
     submit: '提交',
     cancel: '取消',
     reset: '重置',
     publish: '发布',
     close: '关闭',
-    remoteSave: '保存设计',
+    remoteSave: '保存设计并渲染',
   },
   'en-US': {
     save: 'Save',