+++ packages/example-antd-management-fast-framework/src/modelBuilders/formDesign.js
@@ -1,0 +1,78 @@
+import {
+  getTacitlyState,
+  pretreatmentRemoteSingleData,
+  reducerCollection,
+  reducerDefaultParameters,
+  reducerNameCollection,
+} from 'easy-soft-utility';
+
+import { getData, setData } from '../services/formDesign';
+
+export function buildModel() {
+  return {
+    namespace: 'formDesign',
+
+    state: {
+      ...getTacitlyState(),
+    },
+
+    effects: {
+      *get(
+        {
+          payload,
+          alias,
+          pretreatmentSuccessCallback,
+          pretreatmentFailCallback,
+        },
+        { call, put },
+      ) {
+        const response = yield call(getData, payload);
+
+        const dataAdjust = pretreatmentRemoteSingleData({
+          source: response,
+          successCallback: pretreatmentSuccessCallback || null,
+          failCallback: pretreatmentFailCallback || null,
+        });
+
+        yield put({
+          type: reducerNameCollection.reducerRemoteData,
+          payload: dataAdjust,
+          alias,
+          ...reducerDefaultParameters,
+        });
+
+        return dataAdjust;
+      },
+      *set(
+        {
+          payload,
+          alias,
+          pretreatmentSuccessCallback,
+          pretreatmentFailCallback,
+        },
+        { call, put },
+      ) {
+        const response = yield call(setData, payload);
+
+        const dataAdjust = pretreatmentRemoteSingleData({
+          source: response,
+          successCallback: pretreatmentSuccessCallback || null,
+          failCallback: pretreatmentFailCallback || null,
+        });
+
+        yield put({
+          type: reducerNameCollection.reducerRemoteData,
+          payload: dataAdjust,
+          alias,
+          ...reducerDefaultParameters,
+        });
+
+        return dataAdjust;
+      },
+    },
+
+    reducers: {
+      ...reducerCollection,
+    },
+  };
+}