+++ packages/example-antd-management-fast-framework/src/modelBuilders/index.js
@@ -4,6 +4,7 @@ import { buildModel as buildAccessWayModel } from './accessWay';
 import { buildModel as buildCurrentOperatorModel } from './currentAccount';
 import { buildModel as buildCurrentSystemModel } from './currentSystem';
 import { buildModel as buildFlowEditorModel } from './flowEditor';
+import { buildModel as buildFormDesignModel } from './formDesign';
 import { buildModel as buildNoticeModel } from './notice';
 import { buildModel as buildSimpleModel } from './simple';
 import { buildModel as buildTestModel } from './testModel';
@@ -13,6 +14,7 @@ function collectModelBuilder() {
   appendExtraBuilder(buildCurrentOperatorModel);
   appendExtraBuilder(buildCurrentSystemModel);
   appendExtraBuilder(buildFlowEditorModel);
+  appendExtraBuilder(buildFormDesignModel);
   appendExtraBuilder(buildNoticeModel);
   appendExtraBuilder(buildSimpleModel);
   appendExtraBuilder(buildTestModel);