+++ packages/example-antd-management-fast-framework/src/pages/Interaction/DesignView/index.jsx
@@ -1,11 +1,22 @@
+import { Empty } from 'antd';
+
 import { connect } from 'easy-soft-dva';
-import { mergeArrowText, showSimpleInfoMessage } from 'easy-soft-utility';
+import {
+  checkStringIsNullOrWhiteSpace,
+  getValueByKey,
+  mergeArrowText,
+  showSimpleInfoMessage,
+} from 'easy-soft-utility';
 
 import { cardConfig } from 'antd-management-fast-common';
 import {
   buildButton,
   convertOptionOrRadioData,
 } from 'antd-management-fast-component';
+import {
+  DataDisplayer,
+  SchemaDisplayer,
+} from 'antd-management-fast-design-playground';
 
 import { PlaygroundDrawer } from '../../../businessComponents/Drawers/PlaygroundDrawer';
 import { BaseView } from '../BaseView';
@@ -13,7 +24,8 @@ import { code as codeBaseView } from '../BaseView/codeSource';
 
 import { code as codeModalView } from './codeSource';
 
-@connect(({ schedulingControl }) => ({
+@connect(({ formDesign, schedulingControl }) => ({
+  formDesign,
   schedulingControl,
 }))
 class ModalView extends BaseView {
@@ -22,11 +34,16 @@ class ModalView extends BaseView {
 
     this.state = {
       ...this.state,
+      loadApiPath: 'formDesign/get',
       pageTitle: 'Modal 交互示例',
       currentCodeTitle: 'ModalView',
       currentCode: codeModalView,
     };
   }
+
+  afterDesignDrawerClose = () => {
+    this.reloadData({});
+  };
 
   establishToolBarConfig = () => {
     return {
@@ -48,12 +65,71 @@ class ModalView extends BaseView {
   };
 
   establishCardCollectionConfig = () => {
-    const { currentCode, currentCodeTitle } = this.state;
+    const { currentCode, currentCodeTitle, metaData } = this.state;
 
     const that = this;
 
+    const designJson = getValueByKey({
+      data: metaData,
+      key: 'designSchema',
+    });
+
+    const dataSchema = getValueByKey({
+      data: metaData,
+      key: 'dataSchema',
+      defaultValue: '[]',
+    });
+
+    let listDataSchema = JSON.parse(dataSchema);
+
+    const hasDataSchema = listDataSchema.length > 0;
+
+    const designData = {
+      form: {},
+      schema: {},
+      ...(checkStringIsNullOrWhiteSpace(designJson)
+        ? {}
+        : JSON.parse(designJson)),
+    };
+
     return {
       list: [
+        [
+          {
+            title: {
+              text: '表单示例',
+            },
+            width: 'auto',
+            items: [
+              {
+                lg: 24,
+                type: cardConfig.contentItemType.component,
+                component: (
+                  <div>
+                    <SchemaDisplayer {...designData}>
+                      {hasDataSchema ? null : (
+                        <Empty description="暂无表单设计，请进行设计" />
+                      )}
+                    </SchemaDisplayer>
+                  </div>
+                ),
+              },
+            ],
+          },
+          {
+            title: {
+              text: '数据集合',
+            },
+            width: '400px',
+            items: [
+              {
+                lg: 24,
+                type: cardConfig.contentItemType.component,
+                component: <DataDisplayer schema={listDataSchema} />,
+              },
+            ],
+          },
+        ],
         {
           title: {
             text: '代码示例',
@@ -131,7 +207,11 @@ class ModalView extends BaseView {
   renderPresetOther = () => {
     return (
       <>
-        <PlaygroundDrawer />
+        <PlaygroundDrawer
+          afterClose={() => {
+            this.afterDesignDrawerClose();
+          }}
+        />
       </>
     );
   };