+++ packages/antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.11.80
+
+### Patch Changes
+
+- update
+
 ## 2.11.79
 
 ### Patch Changes